feat: Implement board generation for minesweeper
Generate the board by randomly choosing a cell from the board and
place mine on it repeatedly. Note that if the chosen cell already
has a mine on it, another random cell is chosen until no mine is
on the cell, which is quite inefficent, further improvements could
be made.