feat: Implement reveal cell for minesweeper
Implemented `reveal` method. New property `is_revealed` is added for
class `Minesweeper` to determine if the board is being revealed for
the first time. If so, the cell clicked should be guaranteed to not
contain a mine.

While it's possible to regenerate the whole board again when the first
clicked cell contains a mine, it's inefficient to do so. Therefore,
`remove_mine` member function is introduced to remove mine on the grid
and update adjacent cells. Then, we extract the logic of randomly
choosing a cell on the board to place a new mine from `generate_board`
into a new member function `place_mine`.